requirements.txt
torch
torchvision
numpy
matplotlib
tqdm
pyyaml